cooldows and player speed (not working but it have idea)

+++ Shadow Adventure/include/sys/constants.h
@@ -1,3 +1,3 @@
 #pragma once
 
-static int MAX_AMMO = 10;
+static int8_t MAX_AMMO = 10;
